refactor: 지하철 경로 거리 및 요금 계산에서 값의 타입을 double 형으로 수정

+++ src/main/java/wooteco/subway/domain/Fare.java
@@ -2,15 +2,15 @@ package wooteco.subway.domain;
 
 public class Fare {
 
-    private static final int DEFAULT_FARE = 1250;
-    private static final int ADDITIONAL_DISTANCE_PER_5KM = 10;
-    private static final int ADDITIONAL_DISTANCE_PER_8KM = 51;
-    private static final int DISTANCE_UNIT_UNDER_50 = 5;
-    private static final int DISTANCE_UNIT_OVER_50 = 8;
-    private static final int ADDITIONAL_AMOUNT = 100;
+    private static final double DEFAULT_FARE = 1250;
+    private static final double ADDITIONAL_DISTANCE_PER_5KM = 10;
+    private static final double ADDITIONAL_DISTANCE_PER_8KM = 51;
+    private static final double DISTANCE_UNIT_UNDER_50 = 5;
+    private static final double DISTANCE_UNIT_OVER_50 = 8;
+    private static final double ADDITIONAL_AMOUNT = 100;
 
-    public int calculate(final int distance) {
-        int fare = DEFAULT_FARE;
+    public double calculate(final double distance) {
+        double fare = DEFAULT_FARE;
         if (distance >= ADDITIONAL_DISTANCE_PER_5KM && distance < ADDITIONAL_DISTANCE_PER_8KM) {
             return fare + addExtraFare(distance, DISTANCE_UNIT_UNDER_50, ADDITIONAL_DISTANCE_PER_5KM);
         }
@@ -22,7 +22,7 @@ public class Fare {
         return fare;
     }
 
-    private int addExtraFare(final int distance, final int distanceUnit, final int limit) {
-        return (int) Math.ceil((double) (distance - limit) / distanceUnit) * ADDITIONAL_AMOUNT;
+    private double addExtraFare(final double distance, final double distanceUnit, final double limit) {
+        return Math.ceil((distance - limit) / distanceUnit) * ADDITIONAL_AMOUNT;
     }
 }
